requirements.txt
alembic
bcrypt
fastapi[all]
pre-commit
pydantic-settings
pytest
python-jose[cryptography]
asyncpg
fastapi-users[sqlalchemy]
ruff
sqlalchemy[asyncio]
